feat(es): Add Elasticsearch endpoint, clean project
feat(es): Add Elasticsearch endpoint, clean project
