--- camel-ingester/src/main/java/com/zenika/wikipedia/MySpringBean.java
@@ -1,22 +1,0 @@
-package com.zenika.wikipedia;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-/**
- * A bean that returns a message when you call the {@link #saySomething()} method.
- * <p/>
- * Uses <tt>@Component("myBean")</tt> to register this bean with the name <tt>myBean</tt>
- * that we use in the Camel route to lookup this bean.
- */
-@Component("myBean")
-public class MySpringBean {
-
-    @Value("${greeting}")
-    private String say;
-
-    public String saySomething() {
-        return say;
-    }
-
-}
