--- camel-ingester/src/main/java/com/zenika/wikipedia/MySpringBootApplication.java
@@ -1,16 +1,0 @@
-package com.zenika.wikipedia;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MySpringBootApplication {
-
-  /**
-   * A main method to start this application.
-   */
-  public static void main(String[] args) {
-    SpringApplication.run(MySpringBootApplication.class, args);
-  }
-
-}
