+++ camel-ingester/src/main/resources/application-elasticsearch.properties
@@ -1,0 +1,3 @@
+ingest.endpoint.name=direct:es
+ingest.index.name=wikipedia-fr
+ingest.document.type=doc
