+++ camel-ingester/src/main/resources/application-solr.properties
@@ -1,0 +1,4 @@
+ingest.endpoint.name=seda:solr?size=10000&blockWhenFull=true
+solr.connection.string=localhost:8983/solr
+solr.zk.host=localhost:2181
+solr.collection.name=wikipedia
