+++ camel-ingester/src/main/resources/application.properties
@@ -1,44 +1,17 @@
-
 # the name of Camel
-camel.springboot.name = MyCamel
-
-# what to say
-greeting = Hello World
-
+camel.springboot.name=MyCamel
 # how often to trigger the timer
-timer.period = 2000
-
-# to automatic shutdown the JVM after a period of time
-#camel.springboot.duration-max-seconds=60
-#camel.springboot.duration-max-messages=100
-
-# add for example: &repeatCount=5 to the timer endpoint to make Camel idle
-#camel.springboot.duration-max-idle-seconds=15
-
+timer.period=2000
 # expose actuator endpoint via HTTP
 management.endpoints.web.exposure.include=info,health,camelroutes
-
 # turn on actuator health check
-management.endpoint.health.enabled = true
-
+management.endpoint.health.enabled=true
 # allow to obtain basic information about camel routes (read only mode)
-management.endpoint.camelroutes.enabled = true
-management.endpoint.camelroutes.read-only = true
-
+management.endpoint.camelroutes.enabled=true
+management.endpoint.camelroutes.read-only=true
 # to configure logging levels
 #logging.level.org.springframework = INFO
 #logging.level.org.apache.camel.spring.boot = INFO
 #logging.level.org.apache.camel.impl = DEBUG
 #logging.level.sample.camel = DEBUG
-
-camel.component.stax.enabled = true
-
-#data.input.directory = /projets/tests/documents/json
-data.input.directory = /Users/max/Documents/talks/data/final_data
-
-#ingest.endpoint.name = seda:solr?size=10000&blockWhenFull=true
-solr.connection.string = localhost:8983/solr
-solr.zk.host = localhost:2181
-solr.collection.name = wikipedia
-
-ingest.endpoint.name = direct:es
+camel.component.stax.enabled=true
